Whitespace and rename
