+++ src/CMakeLists.txt
@@ -3,11 +3,11 @@
 # Set up the IDE
 set(MAIN_SRC_FILES CovidSim.cpp Rand.cpp Error.cpp Dist.cpp
   Kernels.cpp Bitmap.cpp SetupModel.cpp CalcInfSusc.cpp Sweep.cpp Update.cpp
-  Param.cpp Person.cpp Direction.cpp InverseCdf.cpp Memory.cpp CLI.cpp Files.cpp Params.cpp)
+  Param.cpp Person.cpp Direction.cpp InverseCdf.cpp Memory.cpp CLI.cpp Files.cpp ReadParams.cpp)
 set(MAIN_HDR_FILES CovidSim.h Rand.h Constants.h Country.h Error.h
   Dist.h Kernels.h Bitmap.h Model.h Param.h SetupModel.h ModelMacros.h
   InfStat.h CalcInfSusc.h Sweep.h Update.h MicroCellPosition.hpp Direction.hpp
-  InverseCdf.h Memory.h CLI.h Files.h Params.h)
+  InverseCdf.h Memory.h CLI.h Files.h ReadParams.h)
 source_group(covidsim\\main FILES ${MAIN_SRC_FILES} ${MAIN_HDR_FILES})
 
 # CovidSim target
